Enable Postgres for CI GitHub Action
This runs [the `postgres` image from Docker Hub] as a [service] on the
test machine, which means it communicates via a network bridge. Map 5432
on the Docker instance to 5432 on the host, and export it via the
[DATABASE_URL] for Rails.

[the `postgres` image from Docker Hub]: https://hub.docker.com/_/postgres
[service]: https://docs.github.com/en/actions/guides/about-service-containers
[DATABASE_URL]: https://guides.rubyonrails.org/configuring.html#configuring-a-database

Set the `--health-...` options so GitHub Actions can perform liveness
checks.

More reading material:

- https://docs.github.com/en/actions/guides/creating-postgresql-service-containers
- https://github.com/actions/example-services/blob/master/.github/workflows/postgres-service.yml